# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "PockitTV Online Store",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/store/api/v1/upload/csv": {
            "post": {
                "description": "Upload a CSV file to the server",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CSV Upload"
                ],
                "summary": "Upload a CSV file",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Store ID",
                        "name": "store_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "CSV file to upload",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "File uploaded successfully",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}